PYTHON/requirements.txt
# Discord bot
discord.py==2.3.2

# HTTP requests
aiohttp==3.9.1

# Fuzzy string matching (for boss name detection)
python-Levenshtein==0.23.0

# Environment variables
python-dotenv==1.0.0

# JSON handling (built-in, but listed for clarity)
# json - built-in
# asyncio - built-in